add: test ring buffer is moving

+++ deployment/Core/Inc/ring_buffer.h
@@ -34,7 +34,7 @@ void ring_buffer_get_min(ai_float* min_acc, ai_float* min_gyro);
 void ring_buffer_get_max(ai_float* max_acc, ai_float* max_gyro);
 void ring_buffer_estimate_velocity(int32_t new_samples, float dt);
 void ring_buffer_estimate_gravity(int32_t new_samples);
-int32_t ring_buffer_is_moving(int32_t new_samples);
+int8_t ring_buffer_is_moving(int32_t new_samples);
 float * ring_buffer_get_gravity();
 float * ring_buffer_get_velocity();
 void ring_buffer_reset();
